refac(zk): move `CompressExpressions()` to `zk::plonk` namespace
