+++ tachyon/zk/lookup/halo2/BUILD.bazel
@@ -3,16 +3,6 @@ load("//bazel:tachyon_cc.bzl", "tachyon_cc_library", "tachyon_cc_unittest")
 package(default_visibility = ["//visibility:public"])
 
 tachyon_cc_library(
-    name = "compress_expression",
-    hdrs = ["compress_expression.h"],
-    deps = [
-        "//tachyon/base:compiler_specific",
-        "//tachyon/base:openmp_util",
-        "//tachyon/zk/plonk/expressions:proving_evaluator",
-    ],
-)
-
-tachyon_cc_library(
     name = "evaluator",
     hdrs = ["evaluator.h"],
     deps = [
@@ -44,7 +34,6 @@ tachyon_cc_library(
         "prover_impl.h",
     ],
     deps = [
-        ":compress_expression",
         ":opening_point_set",
         ":permute_expression_pair",
         "//tachyon/base:ref",
@@ -53,6 +42,7 @@ tachyon_cc_library(
         "//tachyon/zk/base/entities:prover_base",
         "//tachyon/zk/lookup:argument",
         "//tachyon/zk/plonk/base:multi_phase_ref_table",
+        "//tachyon/zk/plonk/expressions:compress_expression",
         "//tachyon/zk/plonk/expressions:proving_evaluator",
         "//tachyon/zk/plonk/permutation:grand_product_argument",
         "@com_google_absl//absl/types:span",
@@ -106,15 +96,10 @@ tachyon_cc_library(
 
 tachyon_cc_unittest(
     name = "halo2_unittests",
-    srcs = [
-        "compress_expression_unittest.cc",
-        "permute_expression_pair_unittest.cc",
-    ],
+    srcs = ["permute_expression_pair_unittest.cc"],
     deps = [
-        ":compress_expression",
         ":permute_expression_pair",
         "//tachyon/base:random",
-        "//tachyon/zk/expressions:expression_factory",
-        "//tachyon/zk/lookup/halo2/test:compress_expression_test",
+        "//tachyon/zk/plonk/halo2:bn254_shplonk_prover_test",
     ],
 )
