+++ tachyon/zk/lookup/log_derivative_halo2/BUILD.bazel
@@ -25,9 +25,9 @@ tachyon_cc_library(
         "//tachyon/crypto/commitments:polynomial_openings",
         "//tachyon/zk/base/entities:prover_base",
         "//tachyon/zk/lookup:argument",
-        "//tachyon/zk/lookup/halo2:compress_expression",
         "//tachyon/zk/lookup/halo2:opening_point_set",
         "//tachyon/zk/plonk/base:multi_phase_ref_table",
+        "//tachyon/zk/plonk/expressions:compress_expression",
         "//tachyon/zk/plonk/expressions:proving_evaluator",
     ],
 )
